% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_wide.R
\name{extract_all_in_wide}
\alias{extract_all_in_wide}
\title{Extracts a list of dataframes in wide formats}
\usage{
extract_all_in_wide(list_of_dfs)
}
\arguments{
\item{list_of_dfs}{List of dataframes with nested variable column}
}
\value{
A list of wide-formatted dataframes for the variables previously a nested list
}
\description{
Extracts a list of dataframes in wide formats
}
