% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variable_lookup.R
\name{extract_variable_lookup}
\alias{extract_variable_lookup}
\title{Generates a lookup table for variables and categories}
\usage{
extract_variable_lookup(log_df)
}
\arguments{
\item{log_df}{A parsed dataframe after reading in file}
}
\value{
Lookup table
}
\description{
Generates a lookup table for variables and categories
}
