% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_wide.R
\name{extract_variables_wide}
\alias{extract_variables_wide}
\title{Extracts a wide-format dataframe for a single splitted dataframe}
\usage{
extract_variables_wide(single_splitted_df)
}
\arguments{
\item{single_splitted_df}{A single dataframe}
}
\value{
A wide dataframe for the variables previously a nested list
}
\description{
Extracts a wide-format dataframe for a single splitted dataframe
}
