% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_log.R
\name{read_in_log}
\alias{read_in_log}
\title{Reads in log file}
\usage{
read_in_log(path)
}
\arguments{
\item{path}{Path to the input file}
}
\value{
a dataframe with categories and variables collected.
}
\description{
Reads in file gotten from runnint pymavdump.py
}
