% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_separate_dfs.R
\name{split_separate_dfs}
\alias{split_separate_dfs}
\title{Reads in raw dataframe and splits to list of different splitted dataframes}
\usage{
split_separate_dfs(log_df)
}
\arguments{
\item{log_df}{A parsed dataframe after reading in file}
}
\value{
List of dataframes for each category
}
\description{
Reads in raw dataframe and splits to list of different splitted dataframes
}
